requirements.txt
fastapi==0.111.0
pandas==2.2.2
pytest==8.2.2
python-dotenv==1.0.1
ruff==0.4.8
uvicorn==0.30.1
